% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{myLogFunction}
\usage{
mylog(x)
}
\arguments{
\item{x}{Description of parameter x.}
}
\value{
Description of the return value.
}
\description{
Description of what the function does.
}
\examples{
my_function(5)  # This is a valid example
my_function(c(1, 2, 3))  # Another example
}
